requirements.txt
asgiref==3.7.2
Django==5.0.2
django-model-utils==4.4.0
django-notifications-hq==1.8.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jsonfield==3.1.0
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
swapper==1.3.0
tzdata==2024.1
uritemplate==4.1.1
